Moved the MockValueResolver unit-tests to their own file/class.
