+++ tests/Our.Umbraco.Ditto.Tests/Models/ComplexModel.cs
@@ -20,9 +20,6 @@
         [TypeConverter(typeof(MockPublishedContentConverter))]
         public IPublishedContent MyPublishedContent { get; set; }
 
-        [MockValue("Mock Property Value")]
-        public string MyResolvedProperty { get; set; }
-
         [AppSetting("MyAppSettingKey")]
         public string MyAppSettingProperty { get; set; }
     }
